requirements.txt
mysqlclient==1.3.9
m3-dbfpy==2.3.0

coloredlogs==5.1.1
progressbar2==3.12.0

